+++ src/hooks/player/useIsCurrentTrack.ts
@@ -1,9 +1,10 @@
-import { useCurrentTrack as useOrpheusTrack } from '@roitium/expo-orpheus'
+import { usePlayerStore } from '@/hooks/stores/usePlayerStore'
+import { useShallow } from 'zustand/react/shallow'
 
 export function useIsCurrentTrack(trackUniqueKey: string) {
-	const { track } = useOrpheusTrack()
-
-	return track?.id === trackUniqueKey
+	return usePlayerStore(
+		useShallow((state) => state.orpheusTrack?.id === trackUniqueKey),
+	)
 }
 
 export default useIsCurrentTrack